Banker's Algorithm_final revision
This Revision in the following repository is the Complete code for the Banker's Algorithm to check the Safety State condition of the given matrix.
Also, the Data fields of Max, Allocated, Available are predefined in the program. However, if are changed or initialized at the run time won't affect the logic of the program.